move abstract class to another file
